+++ packages/cf-backend/tests/unit-alarm-tracing.test.ts
@@ -24,6 +24,7 @@ const PHASES = [
   'alarm.cache_warm',
   'alarm.sleep_time',
   'alarm.timer_rearm',
+  'alarm.slate_pictures',
 ] as const;
 
 describe('alarm tick tracing', () => {
@@ -108,6 +109,7 @@ describe('alarm tick tracing', () => {
         '  alarm.cache_warm  [isolate_gen=1 invocation=1 cache_warmed=false]',
         '  alarm.sleep_time  [isolate_gen=1 invocation=1 sleep_time_ran=false]',
         '  alarm.timer_rearm  [isolate_gen=1 invocation=1 rearmed=false]',
+        '  alarm.slate_pictures  [isolate_gen=1 invocation=1]',
       ].join('\n'),
     );
   });